# docs/gpt-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Teed API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://teed.club"
    }
  ],
  "paths": {
    "/api/gpt/bags": {
      "get": {
        "operationId": "listBags",
        "summary": "List all bags owned by the user",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "operationId": "createBag",
        "summary": "Create a new bag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {"type": "string"},
                  "description": {"type": "string"},
                  "category": {"type": "string"},
                  "is_public": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/gpt/bags/{code}": {
      "get": {
        "operationId": "getBag",
        "summary": "Get a specific bag",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "operationId": "updateBag",
        "summary": "Update a bag",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "description": {"type": "string"},
                  "category": {"type": "string"},
                  "is_public": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "operationId": "deleteBag",
        "summary": "Delete a bag",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/gpt/bags/{code}/items": {
      "get": {
        "operationId": "listItems",
        "summary": "List items in a bag",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "operationId": "addItem",
        "summary": "Add an item to a bag",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string"},
                  "brand": {"type": "string"},
                  "description": {"type": "string"},
                  "category": {"type": "string"},
                  "link": {"type": "string"},
                  "image_url": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/gpt/items/{id}": {
      "patch": {
        "operationId": "updateItem",
        "summary": "Update an item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "brand": {"type": "string"},
                  "description": {"type": "string"},
                  "link": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "operationId": "deleteItem",
        "summary": "Delete an item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/gpt/profile": {
      "get": {
        "operationId": "getProfile",
        "summary": "Get user profile",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}